feat: validateDuplicate(), validateRange() 추가
